+++ Macros/Characters.hpp
