+++ RequestHandler/Messenger/Messenger.hpp
@@ -8,11 +8,11 @@
 
 #include "../../Channel/Channel.hpp"
 #include "../../Client/Client.hpp"
-#include "../../Macros/Character.hpp"
+#include "../../Macros/Characters.hpp"
+#include "../../Macros/Errors.hpp"
+#include "../../Macros/Replies.hpp"
 #include "../../Print/Print.hpp"  // TODO: delete
 #include "../../Server/Server.hpp"
-#include "../../Macros/Error.hpp"
-#include "../../Macros/Reply.hpp"
 
 class Messenger {
  private:
